add policy

+++ app/Http/Controllers/VacancyController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\Vacancy\VacancyStoreRequest;
 use App\Http\Requests\Vacancy\VacancyUpdateRequest;
 use App\Http\Resources\VacancyCollection;
 use App\Http\Resources\VacancyResource;
+use App\Models\Organization;
 use App\Models\Vacancy;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -23,6 +24,7 @@ class VacancyController extends Controller
             if ($only_active != "false") {
                 if ($value->workers_booked < $value->workers_amount) return $value;
             } else {
+                $this->authorize('index', Vacancy::class);
                 return $value;
             }
         });
@@ -31,23 +33,34 @@ class VacancyController extends Controller
 
     public function store(VacancyStoreRequest $request)
     {
-        $vacancy = Vacancy::create($request->all());
+        $data = $request->all();
+        $organization = Organization::where('id', $data['organization_id'])->first();
+
+        if ($organization->user_id != Auth::guard('api')->id()) $this->authorize('create', Vacancy::class);
+
+        $vacancy = Vacancy::create($data);
 
         return new VacancyResource($vacancy);
     }
 
     public function show(Vacancy $vacancy)
     {
+
         $vacancy->load(['workers']);
 
         $vacancy->workers_booked = count($vacancy->workers()->get());
-        $vacancy->_workers = true;
+        if (Auth::guard('api')->user()->role !== 'worker') {
+            // $this->authorize('show', $vacancy);
+            $vacancy->_workers = true;
+        }
 
         return new VacancyResource($vacancy);
     }
 
     public function update(VacancyUpdateRequest $request, Vacancy $vacancy)
     {
+        $this->authorize('update', $vacancy);
+
         $vacancy->update($request->all());
 
         return new VacancyResource($vacancy);
@@ -55,6 +68,8 @@ class VacancyController extends Controller
 
     public function destroy(Vacancy $vacancy)
     {
+        $this->authorize('delete', $vacancy);
+
         $vacancy->delete();
 
         return new VacancyResource($vacancy);
@@ -66,7 +81,7 @@ class VacancyController extends Controller
         $user_id = $request->post('user_id');
         $vacancy_id = $request->post('vacancy_id');
 
-        if ($id !== $user_id) abort(404);
+        if ($id !== $user_id) $this->authorize('book', Vacancy::class);
 
         $vacancy = Vacancy::find($vacancy_id);
         $vacancy->workers()->attach($user_id);
@@ -82,8 +97,8 @@ class VacancyController extends Controller
         $user_id = $request->post('user_id');
         $vacancy_id = $request->post('vacancy_id');
 
-        if ($id !== $user_id) abort(404, 'error');
-
+        if ($id !== $user_id) $this->authorize('book', Vacancy::class);
+        
         $vacancy = Vacancy::find($vacancy_id);
         $vacancy->workers()->detach($user_id);
